+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+* add Best Buy stock monitoring with watch list support
+
 ## v0.6.0 (2026-01-05)
 
 ## [0.6.0](https://github.com/ltruong0/marcellobot/compare/v0.5.0...v0.6.0) (2026-01-05)